feat: convert Form > Base Elements page to TS
