--- showcase/app/routes/page-components/form/base-elements.js
@@ -1,19 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: MPL-2.0
- */
-
-import Route from '@ember/routing/route';
-
-export default class PageComponentsFormBaseElementsRoute extends Route {
-  model() {
-    // these are used only for presentation purpose in the showcase
-    const SAMPLE_ERROR_MESSAGES = [
-      'First error message',
-      'Second error message',
-    ];
-    return {
-      SAMPLE_ERROR_MESSAGES,
-    };
-  }
-}
